change code to ES6 syntax
