+++ index.html
@@ -7,7 +7,7 @@
     <!-- App title -->
     <title>Awosome Book with navigation</title>
     <!-- Link CSS file -->
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="./styles/index.css">
 </head>
 <body>
     <!-- Navigation bar -->
@@ -65,7 +65,7 @@
       <p>You can reach out to us!</p>
       <ul>
         <li><a href="#"></a>ndikumanaisaie@gmail.com</li>
-        <li><a href="#"></a>ndikumanaisaie@gmail.com</li>
+        <li><a href="#"></a>ndikumanaisaie@yahoo.com</li>
         <li><a href="#"></a>kinyeramo@gmail.com</li>
       </ul>
     </section>
@@ -74,6 +74,6 @@
       <p>Copyright</p>
     </footer>
     
-    <script src="main.js"></script>
+    <script type="module" src="./modules/index.js"></script>
 </body>
 </html>
